migrate to Framer Motion
