+++ package.json
@@ -23,6 +23,7 @@
     "@giphy/js-fetch-api": "^4.1.2",
     "axios": "^0.21.4",
     "file-saver": "^2.0.5",
+    "framer-motion": "^4.1.17",
     "gatsby": "^3.13.0",
     "gatsby-plugin-image": "^1.13.0",
     "gatsby-plugin-manifest": "^3.13.0",
@@ -43,7 +44,6 @@
     "react-dom": "^17.0.2",
     "react-feather": "^2.0.9",
     "react-helmet": "^6.1.0",
-    "react-pose": "^4.0.10",
     "remark": "^13.0.0",
     "remark-breaks": "^2.0.2",
     "remark-html": "^13.0.2",
